Entregas - Victor Alexandre/ADK_Agents/requirements.txt
google-adk
google-genai
python-dotenv
